refactoring class packages
