+++ src/main/java/com/mageddo/dnsproxyserver/docker/DockerNetworkService.java
@@ -12,7 +12,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static com.mageddo.dnsproxyserver.docker.ContainerSolvingService.NETWORK_MODE_HOST;
+import static com.mageddo.dnsproxyserver.server.dns.solver.docker.application.ContainerSolvingService.NETWORK_MODE_HOST;
 import static com.mageddo.dnsproxyserver.docker.DpsContainerManager.isDpsContainer;
 
 @Slf4j
