+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/application/ContainerSolvingService.java
@@ -1,7 +1,12 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.application;
 
 import com.github.dockerjava.api.command.InspectContainerResponse;
 import com.mageddo.dnsproxyserver.config.Configs;
+import com.mageddo.dnsproxyserver.docker.DockerDAO;
+import com.mageddo.dnsproxyserver.docker.DockerNetworkDAO;
+import com.mageddo.dnsproxyserver.docker.Entry;
+import com.mageddo.dnsproxyserver.docker.Labels;
+import com.mageddo.dnsproxyserver.docker.NetworkComparator;
 import com.mageddo.dnsproxyserver.server.dns.solver.HostnameQuery;
 import com.mageddo.net.IP;
 import com.mageddo.net.Networks;
