+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/solver/docker/application/MatchingContainerService.java
@@ -1,6 +1,7 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.server.dns.solver.docker.application;
 
 import com.github.dockerjava.api.command.InspectContainerResponse;
+import com.mageddo.dnsproxyserver.docker.DockerDAO;
 import com.mageddo.dnsproxyserver.server.dns.solver.HostnameQuery;
 import lombok.RequiredArgsConstructor;
