+++ src/test/java/com/mageddo/dnsproxyserver/docker/ContainerHostnameMatcherTest.java
@@ -2,6 +2,7 @@ package com.mageddo.dnsproxyserver.docker;
 
 import com.mageddo.dnsproxyserver.config.Configs;
 import com.mageddo.dnsproxyserver.server.dns.solver.HostnameQuery;
+import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.ContainerHostnameMatcher;
 import testing.templates.HostnameQueryTemplates;
 import testing.templates.HostnameTemplates;
 import testing.templates.docker.InspectContainerResponseTemplates;
